backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
websockets==12.0
httpx==0.27.0
python-dotenv==1.0.1

# Database
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
alembic==1.13.1

# Caching
redis[asyncio]==5.0.4

# Scheduling
apscheduler==3.10.4

# AI / ML
google-generativeai==0.7.2
stable-baselines3==2.3.2
torch==2.3.0
gymnasium==0.29.1
# torch-geometric — install manually with correct CUDA/CPU wheel
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.4.2

# Vector Store / RAG
chromadb==0.5.3

# Broker SDKs
oandapyV20==0.7.2
alpaca-py==0.21.1
python-binance==1.0.19

# News / RSS
feedparser==6.0.11
requests==2.31.0
beautifulsoup4==4.12.3

# Finance / Risk
scipy==1.13.0

# PayPal
paypalrestsdk==1.13.3

# Testing
pytest==8.2.2
pytest-asyncio==0.23.7
httpx==0.27.0
